rag_service/app/requirements.txt
streamlit
langchain
qdrant-client
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
python-dotenv